requirements.txt
beautifulsoup4==4.5.1
chardet==3.0.4
click==6.6
dimensions==0.0.2
futures==3.0.5
futures-then==0.1.1
humanfriendly==1.44.7
jinja2==2.10
langdetect==1.0.6
lxml>=3.6.4
Markdown==2.6.6
painter==0.3.1
py==1.6.0
pytest==3.8.2
regex>=2017.6.23
six==1.10.0
python-magic>=0.4.15
ipaddress==1.0.22
jsoncomment==0.3.0
